mlkem: define Montgomery u via uint32_t product (fix AppleClang ML-KEM)
Computing u as (int16_t)(a * MLKEM_QINV) uses signed int multiply; for typical
int32 intermediates |a|*|QINV| exceeds INT_MAX — undefined behavior. GCC on
Linux often matched the intended low-16-bit modular math; Clang in Release on
macOS could mis-compile NTT scalar butterflies, breaking shared-secret tests
(job 420787 after prior fixes).

Use (uint32_t)a * (uint16_t)QINV for u per pq-crystals-style ML-KEM refs.
Deduplicate MLKEM _basemul steps onto dap_mlkem_montgomery_reduce.

Made-with: Cursor

+++ module/crypto/src/kem/mlkem/dap_mlkem_reduce.h
@@ -11,7 +11,8 @@
 
 static inline int16_t dap_mlkem_montgomery_reduce(int32_t a)
 {
-    int16_t u = (int16_t)(a * MLKEM_QINV);
+    /* uint32 path: signed int32 * int16 exceeds INT_MAX — UB with plain int multiply (Clang vs GCC). */
+    int16_t u = (int16_t)((uint32_t)a * (uint32_t)(uint16_t)MLKEM_QINV);
     return (int16_t)((a - (int32_t)u * MLKEM_Q) >> 16);
 }
